Added warning when using WaterLily in single thread mode (JULIA_NUM_THREADS=1). Added allocations test in the pipeline. Conditional loading of CUDA and AMDGPU backends on the pipeline to avoid huge error message.
