+++ src/WaterLily.jl
@@ -132,6 +132,18 @@ function restart_sim! end
 # export
 export restart_sim!
 
+# Check number of threads when loading WaterLily
+"""
+    check_nthreads(::Val{1})
+
+Check the number of threads available for the Julia session that loads WaterLily.
+A warning is shown when running in serial (JULIA_NUM_THREADS=1).
+"""
+check_nthreads(::Val{1}) = @warn("\nUsing WaterLily in serial (ie. JULIA_NUM_THREADS=1) is not recommended because \
+    it disables the GPU backend and defaults to serial CPU."*
+    "\nUse JULIA_NUM_THREADS=auto, or any number of threads greater than 1, to allow multi-threading in CPU or GPU backends.")
+check_nthreads(_) = nothing
+
 # Backward compatibility for extensions
 if !isdefined(Base, :get_extension)
     using Requires
@@ -143,6 +155,7 @@ function __init__()
         @require WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192" include("../ext/WaterLilyWriteVTKExt.jl")
         @require ReadVTK = "dc215faf-f008-4882-a9f7-a79a826fadc3" include("../ext/WaterLilyReadVTKExt.jl")
     end
+    check_nthreads(Val{Threads.nthreads()}())
 end
 
 end # module
